(#96) Volta do modulo js do popup
