--- src/CHART/teste.html
@@ -1,68 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <script src="https://cdn.jsdelivr.net/npm/chart.js@3.1.0/dist/chart.min.js"></script>
-</head>
-<body>
-    <canvas id="myChart" width="400" height="400"></canvas>
-    <a id="download"
-    download="ChartImage.jpg" 
-    href=""
-    class="btn btn-primary float-right bg-flat-color-1"
-    title="Descargar Gráfico">TESTE DE GERAR PNG
-
-    <!-- Download Icon -->
-<i class="fa fa-download"></i>
-</a>
-    <script>
-        var ctx = document.getElementById('myChart').getContext('2d');
-        var myChart = new Chart(ctx, {
-            type: 'bar',
-            data: {
-                labels: ['Red', 'Blue', 'Yellow', 'Green', 'Purple', 'Orange'],
-                datasets: [{
-                    label: '# of Votes',
-                    data: [12, 19, 3, 5, 2, 3],
-                    backgroundColor: [
-                        'rgba(255, 99, 132, 0.2)',
-                        'rgba(54, 162, 235, 0.2)',
-                        'rgba(255, 206, 86, 0.2)',
-                        'rgba(75, 192, 192, 0.2)',
-                        'rgba(153, 102, 255, 0.2)',
-                        'rgba(255, 159, 64, 0.2)'
-                    ],
-                    borderColor: [
-                        'rgba(255, 99, 132, 1)',
-                        'rgba(54, 162, 235, 1)',
-                        'rgba(255, 206, 86, 1)',
-                        'rgba(75, 192, 192, 1)',
-                        'rgba(153, 102, 255, 1)',
-                        'rgba(255, 159, 64, 1)'
-                    ],
-                    borderWidth: 1
-                }]
-            },
-            options: {
-                scales: {
-                    y: {
-                        beginAtZero: true
-                    }
-                }
-            }
-        });
-        //Download Chart Image
-document.getElementById("download").addEventListener('click', function(){
-  /*Get image of canvas element*/
-  var url_base64jp = document.getElementById("lineChart").toDataURL("image/jpg");
-  /*get download button (tag: <a></a>) */
-  var a =  document.getElementById("download");
-  /*insert chart image url to download button (tag: <a></a>) */
-  a.href = url_base64jp;
-});
-        </script>
-</body>
-</html>
