+++ src/index.js
@@ -44,13 +44,61 @@ document.addEventListener("click", function (e) {
     CodigoDevicePost();
   }
 
+  import Chart from 'chart.js';
+function criaGrafico(todosCommits){
+  var myChart = new Chart(ctx, {
+    type: 'bar',
+    data: {
+        labels: ['Red', 'Blue', 'Yellow', 'Green', 'Purple', 'Orange'],
+         datasets: [{
+            label: '# of Votes',
+                data: [todosCommits, 19, 3, 5, 2, 3],
+                backgroundColor: [
+                    'rgba(255, 99, 132, 0.2)',
+                    'rgba(54, 162, 235, 0.2)',
+                    'rgba(255, 206, 86, 0.2)',
+                    'rgba(75, 192, 192, 0.2)',
+                    'rgba(153, 102, 255, 0.2)',
+                    'rgba(255, 159, 64, 0.2)'
+                ],
+                borderColor: [
+                    'rgba(255, 99, 132, 1)',
+                    'rgba(54, 162, 235, 1)',
+                    'rgba(255, 206, 86, 1)',
+                    'rgba(75, 192, 192, 1)',
+                    'rgba(153, 102, 255, 1)',
+                    'rgba(255, 159, 64, 1)'
+                ],
+                borderWidth: 1
+            }]
+        },
+        options: {
+            scales: {
+                y: {
+                    beginAtZero: true
+                }
+            }
+        }
+    });
+    //Download Chart Image
+function dowloadImagem(){
+    /*Get image of canvas element*/
+    var url_base64jp = document.getElementById("lineChart").toDataURL("image/jpg");
+    /*get download button (tag: <a></a>) */
+    var a =  document.getElementById("download");
+    /*insert chart image url to download button (tag: <a></a>) */
+    a.href = url_base64jp;
+
+}
+}
   if(urlParts[2] == "github.com" && urlParts[3] == "login" && urlParts[4] == "device" && urlParts[5] == "success" ){
    
     ConfirmaLoginContaUsuario();
     
-    contribuinteRepositorio(imprime);
+    contribuinteRepositorio(criaGrafico());
     
   }
+
