+++ src/popup/popup.html
@@ -1,13 +1,12 @@
 <html>
   <head>
     <meta charset="UTF-8">
-    <link href="popup.css" rel="stylesheet">
+    
+    <script src="https://cdn.jsdelivr.net/npm/chart.js@3.1.0/dist/chart.min.js"></script>
   </head>
 
   <body>
-    <h1 class="text title">Seja bem vindo ao PullBot!</h1>
-    <p class="text">Somos uma extensão para o Google Chrome que tem por finalidade te auxiliar no monitoramento de produtividade de equipes de desenvolvimento que utilizam o GitHub como plataforma de hospedagem de projetos.</p>
-    
-    <p class="text">Caso existam dúvidas a respeito da utilização, funcionalidades, formas de contribuição do projeto ou qualquer outra coisa, acesse nosso <a href="https://github.com/fga-eps-mds/PullBot" target="_blank">repositório.</a></p>
+    <canvas id="myChart" ></canvas>
+   <script src="popup.js"></script>
   </body>
 </html>
